% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree_for_treejack.R
\name{all_path_levels}
\alias{all_path_levels}
\title{All Paths to Parent Folders}
\usage{
all_path_levels(path)
}
\arguments{
\item{path}{one character string representing a file path}
}
\value{
vector of character representing all parent paths and the \code{path}
  itself as the last element
}
\description{
For a given path \code{a/b/c/d}, all the parent paths \code{a}, \code{a/b},
\code{a/b/c} and the path itself (\code{a/b/c/d}) are returned.
}
\examples{
paths <- kwb.fakin:::all_path_levels("this/is/a/long/path")
kwb.file:::to_subdir_matrix(paths)

}
