% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{createLocalProject}
\alias{createLocalProject}
\title{Create Empty Project Folder Structure Locally}
\usage{
createLocalProject(project)
}
\arguments{
\item{project}{Project name. Must correspond with the name of a folder below
one of the server locations returned by \code{\link{getProjectPaths}}}
}
\description{
Copy the pure folder structure (without files) of a given KWB project from
the server to a new local folder below
"C:/Users/<user_name>/Documents/Projekte"
}
