% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_properties.R
\name{cut.path_tree}
\alias{cut.path_tree}
\title{Cut a Path Tree}
\usage{
\method{cut}{path_tree}(tree, n_levels = 2, depth = 0)
}
\arguments{
\item{tree}{object of class \code{tree} as returned by
\code{kwb.fakin:::to_tree}}

\item{n_levels}{number of levels to which the tree is cut}

\item{depth}{current depth level}
}
\description{
Reduce a path tree to its first levels.
}
