% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extdata_file}
\alias{extdata_file}
\title{Path to File in extdata Folder of this Package}
\usage{
extdata_file(file)
}
\arguments{
\item{file}{filename or path to the file, relative to extdata/}
}
\description{
Path to File in extdata Folder of this Package
}
