% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_properties.R
\name{extract_properties}
\alias{extract_properties}
\title{Extract Pattern-Defined Properties from Strings}
\usage{
extract_properties(x, patterns, replacements, as_data_frame = FALSE)
}
\arguments{
\item{x}{vector of character}

\item{patterns}{vector of character}

\item{replacements}{vector of character}

\item{as_data_frame}{logical. If \code{TRUE} (the default is \code{FALSE}),
a data frame is returned.}
}
\description{
Extract Pattern-Defined Properties from Strings
}
\examples{
# Define patterns to be matched against
patterns <- c(
  "[Bb]ericht",
  "[- ](\\\\d+)$",
  "Abschluss",
  "Zwischen",
  "_HS$"
)

# Define property:value pairs (or even prop1:value1+prob2:value2+...)
# referring to parts of the pattern enclosed in parentheses with \\1, \\2, ...
replacements <- c(
  "type:report",
  "number:\\\\1",
  "stage:final",
  "stage:intermediate",
  "author:Sonnenberg+reviewed:true"
)

# Define strings in which to look for properties and their values
x <- c("Bericht", "Bericht 1", "Abschlussbericht", "Zwischenbericht_HS")

# Extract property values as strings
extract_properties(x = x, patterns, replacements)

# Arrange the properties in a data frame
extract_properties(x = x, patterns, replacements, as_data_frame = TRUE)
}
