% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_file_info.R
\name{get_and_save_file_info}
\alias{get_and_save_file_info}
\title{Get and Save File Information}
\usage{
get_and_save_file_info(root_dir, output_dir)
}
\arguments{
\item{root_dir}{path to the directory from which to start searching for
files}

\item{output_dir}{path to the output directory. In this directory, a file
"path-info_<date-time>_<parent-folder>.csv" will be generated with
<date-time> being a date and time string in yyyy-mm-dd_HHMM format and
<parent-folder> being the last path segment of \code{root_dir}}
}
\description{
Get and Save File Information
}
