% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_permissions.R
\name{get_example_read_and_write_matrices}
\alias{get_example_read_and_write_matrices}
\title{Get Example Matrices of Read- and Write-Permissions}
\usage{
get_example_read_and_write_matrices()
}
\description{
Get Example Matrices of Read- and Write-Permissions
}
