% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_file_duplicates}
\alias{get_file_duplicates}
\title{Get list of paths containing files of the same name}
\usage{
get_file_duplicates(paths, pattern, ...)
}
\arguments{
\item{paths}{vector of character representing full file paths}

\item{pattern}{\code{paths} is matched against this pattern before the
matching paths are split and analysed for duplicated files}

\item{...}{arguments passed to \code{grep}, e.g. \code{ignore.case}}
}
\description{
Get list of paths containing files of the same name
}
\examples{
paths <- c("/a/b/c.exe", "/a/b/d.exe", "/A/B/c.exe", "/A/B/d.exe")
get_file_duplicates(paths, pattern = "\\\\.exe$")

}
