% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_usage.R
\name{get_function_call_frequency}
\alias{get_function_call_frequency}
\title{Which Function is Called How Often?}
\usage{
get_function_call_frequency(tree)
}
\arguments{
\item{tree}{parse tree as returned by \code{\link{parse_scripts}}}
}
\value{
data frame with columns \code{name} (name of function), \code{count}
  (number of times the function is called)
}
\description{
Which Function is Called How Often?
}
