% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names_of_used_packages.R
\name{get_names_of_used_packages}
\alias{get_names_of_used_packages}
\title{Get Names of Packages Used in R-Scripts}
\usage{
get_names_of_used_packages(root_dir, pattern = "[.][rR](md)?$")
}
\arguments{
\item{root_dir}{directory in which to look recursively for R-scripts}

\item{pattern}{regular expression matching the names of the files to be
considered}
}
\description{
Get Names of Packages Used in R-Scripts
}
