% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_path_stat_matrix}
\alias{get_path_stat_matrix}
\title{How Many Folders of the Template are in the Project Folder?}
\usage{
get_path_stat_matrix(project_folder, template_folders)
}
\arguments{
\item{project_folder}{folder in which to look for the projects of one
KWB department, e.g. ".../SUW_Department/Projects"}

\item{template_folders}{vector of relative paths of folders expected to
be contained within each folder within \code{project_folder}}
}
\description{
How Many Folders of the Template are in the Project Folder?
}
