% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recursive_file_info.R
\name{get_recursive_file_info}
\alias{get_recursive_file_info}
\title{Call file.info recursively on files below a root folder}
\usage{
get_recursive_file_info(root_dir, pattern = NULL, all = TRUE,
  dbg = TRUE)
}
\arguments{
\item{root_dir}{path to the root directory from which to start the recursive
search for files}

\item{pattern}{regular expression matching the names of the files to be
considered. By default, all files are considered.}

\item{all}{if \code{TRUE} (default) hidden files are also returned}

\item{dbg}{if \code{TRUE} (default) progress messages are shown}
}
\description{
Call file.info recursively on files below a root folder
}
