% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialChars.R
\name{get_special_character_info}
\alias{get_special_character_info}
\title{Get Special Characters and Their Byte Codes}
\usage{
get_special_character_info(text)
}
\arguments{
\item{text}{vector of character of length one}
}
\value{
data frame with columns \code{special} (containing the special
  characters) and \code{bytes} (containing the hexadecimanl byte codes as
  as space separated string)
}
\description{
Get Special Characters and Their Byte Codes
}
\examples{
(text <- kwb.fakin:::example_string_with_specials("de"))

get_special_character_info(text)

}
