% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{getDictOneByOne}
\alias{getDictOneByOne}
\title{Get a Path Dictionary}
\usage{
getDictOneByOne(paths, n = 10)
}
\arguments{
\item{paths}{vector of character strings representing file or folder paths}

\item{n}{number of compression levels}
}
\description{
Get a Path Dictionary
}
