% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getElementLengths}
\alias{getElementLengths}
\title{Get the Lenghts of List Elements}
\usage{
getElementLengths(x)
}
\arguments{
\item{x}{a list}
}
\value{
vector of integer
}
\description{
Get the Lenghts of List Elements
}
