% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{getProjectPaths}
\alias{getProjectPaths}
\title{Paths where to find project on the KWB server}
\usage{
getProjectPaths(startDirectory)
}
\arguments{
\item{startDirectory}{Path to the project network drive at KWB}
}
\description{
Paths where to find project on the KWB server
}
