% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_file_metadata.R
\name{guess_file_metadata}
\alias{guess_file_metadata}
\title{Guess Metadata about a Text File}
\usage{
guess_file_metadata(file, n_first_rows = 1000, ...)
}
\arguments{
\item{file}{path to text file}

\item{n_first_rows}{number of first rows of \code{file} from which to guess
the meta information.}

\item{\dots}{further arguments passe to \code{\link{read_lines}}, such as
\code{fileEncoding}}
}
\value{
data frame with columns
\itemize{
  \item \code{paths}: does the file seem to contain path information, i.e.
  were slashes or backslashes found?
  \item \code{forbidden}: does the file contain characters that are forbidden
  in file paths?
  \item \code{header}: does the file seem to contain a header row?
  \item \code{windows}: are the paths given in "windows"-style, i.e. are the
    path segments separated by backslash?
  \item \code{sep}: column separator guessed
  \item \code{ncol}: number of columns guessed
} and attributes
\itemize{
  \item \code{file}: a copy of the file path given in \code{file}
  \item \code{first_rows}: first \code{n_first_rows} rows of \code{file}
  \item \code{columns} (optional): column headers if the file is assumed to
  contain a header row
}
}
\description{
Guess Metadata about a Text File
}
