% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{left_substring_equals}
\alias{left_substring_equals}
\title{Is Left Substring of X Equal To Y?}
\usage{
left_substring_equals(x, y)
}
\arguments{
\item{x}{String of which the left part is compared with \code{y}}

\item{y}{String to be compared with the left part of \code{x}}
}
\description{
Is Left Substring of X Equal To Y?
}
