% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{list_files}
\alias{list_files}
\title{Get File Paths Recursively (Windows only!)}
\usage{
list_files(root, file, use_batch = TRUE)
}
\arguments{
\item{root}{path to the directory from which to start "downwards" and
recursively for files and folders.}

\item{file}{path to the result file (text) to which the paths are to be
written.}

\item{use_batch}{if \code{TRUE} (default), a batch file is written (by
default to \code{list_files.bat} in \code{tempdir()}) and run to perform
the dir command}
}
\description{
Get a full list of files below a \code{root} directory using the \code{dir}
function and write it to a given \code{file}
}
