% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_file-name.R
\name{name_is_ok}
\alias{name_is_ok}
\title{Is the Name Ok According to Our Best Practices?}
\usage{
name_is_ok(x, mildness = 1)
}
\arguments{
\item{x}{vector of character}

\item{mildness}{level of mildness. 1: not mild, all characters must be
hyphen or alphanumeric or dot or underscore, 2: more mild, all characters
must be one of the above or space}
}
\value{
vector of logical as long as \code{x}
}
\description{
Is the Name Ok According to Our Best Practices?
}
\examples{
name_is_ok(c("a", "$", ".", " "))
name_is_ok(c("a", "$", ".", " "), mildness = 2)

}
