% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path-tree.R
\name{name_to_traffic_light}
\alias{name_to_traffic_light}
\title{Get Traffic Light Colours for Names}
\usage{
name_to_traffic_light(x)
}
\arguments{
\item{x}{character of (file or folder) names, e.g. as they appear as
node labels in the plot generated with \code{\link{plot_path_network}}}
}
\value{
vector of colour strings each of which is \code{green} (name does
  comply with naming convention), \code{yellow} (name does almost comply with
  naming convention), \code{red} (name does not comply with naming
  convention).
}
\description{
Get Traffic Light Colours for Names
}
\examples{
# Define a vector of names
x <- c("has_speci&l", "has space", "is_ok")

# Colour names by their compliance with naming convention
name_to_traffic_light(x)

}
