% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path-tree.R
\name{names_to_colours_good_name}
\alias{names_to_colours_good_name}
\title{Get Traffic Light Colours for Names}
\usage{
names_to_colours_good_name(node_names)
}
\value{
vector of colour strings each of which is \code{green} (name does
  comply with naming convention), \code{yellow} (name does almost comply with
  naming convention), \code{red} (name does not comply with naming
  convention).
}
\description{
Get Traffic Light Colours for Names
}
\examples{
# Define a vector of names
node_names <- c("has_speci@l", "has space", "is_ok")

# Colour names by their compliance with naming convention
kwb.fakin:::names_to_colours_good_name(node_names)

}
