% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treemaps.R
\name{plot_all_treemaps}
\alias{plot_all_treemaps}
\title{Plot Treemaps for All given Path Infos}
\usage{
plot_all_treemaps(path_infos, as_png = TRUE, ...)
}
\arguments{
\item{path_infos}{list of data frames each of which contains file path
information as returned by \code{\link{read_file_info}}}

\item{as_png}{if \code{TRUE} the plots are saved to png-files in
\code{tempdir()}. The name is then taken from the names of the elements in
\code{path_infos}. Otherwise the plot go into the current graphical device.}

\item{\dots}{further arguments passed to
\code{\link{plot_treemaps_from_path_data}}, such as \code{n_levels}}
}
\value{
for \code{as_png = TRUE} vector of paths to the created png files.
}
\description{
Plot Treemaps for All given Path Infos
}
