% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_biggest_folders}
\alias{plot_biggest_folders}
\title{Plot Folders with Maximum Number of Files}
\usage{
plot_biggest_folders(tree, max_depth = 5, to_pdf = TRUE)
}
\arguments{
\item{tree}{tree list structure as returned by \code{\link{to_tree}}}

\item{max_depth}{maximum number of folder depth}

\item{to_pdf}{if \code{TRUE} the output is directed to a temporary pdf file}
}
\description{
Plot Folders with Maximum Number of Files
}
