% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_file_distribution.R
\name{plot_file_distribution}
\alias{plot_file_distribution}
\title{Plot Distributions of Files in Folder Depths and Name Summaries}
\usage{
plot_file_distribution(file_data, start_path, n_root_parts, ..., to_pdf = TRUE)
}
\arguments{
\item{file_data}{data frame with columns \code{path}, \code{type},
\code{size}}

\item{start_path}{only paths from \code{file_data} are considered that start
with this path}

\item{n_root_parts}{number of first path segments to be considered as "root"}

\item{\dots}{further arguments passed to
\code{kwb.fakin:::plot_file_size_in_depth}}

\item{to_pdf}{if \code{TRUE} (default) the results are plotted into a
temporary pdf file that is opened in a pdf viewer}
}
\description{
Plot Distributions of Files in Folder Depths and Name Summaries
}
\examples{
# Set a seed for the random number generator
set.seed(20190625)

# Create random paths
paths <- kwb.pathdict::random_paths()

# Number of paths
n <- length(paths)

# Create artificial file data (invent sizes)
file_data <- kwb.utils::noFactorDataFrame(
  path = paths,
  type = "file",
  size = abs(rnorm(n)) * kwb.fakin:::bytes_to_mib(
    2^sample(30, n, replace = TRUE)
  )
)

kwb.fakin::plot_file_distribution(
  file_data, start_path = "reason", n_root_parts = 1, to_pdf = FALSE
)

}
