% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-files-in-depth.R
\name{plot_files_in_depth}
\alias{plot_files_in_depth}
\title{Plot File Sizes over Folder Depth}
\usage{
plot_files_in_depth(depth_summaries, project)
}
\arguments{
\item{depth_summaries}{list as returned by \code{\link{get_depth_summaries}}}

\item{project}{name of project (in fact folder name in folder depth 1)}
}
\description{
Plot File Sizes over Folder Depth
}
