% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_number_of_elements_per_folder.R
\name{plot_number_of_elements_per_folder}
\alias{plot_number_of_elements_per_folder}
\title{Plot Folders with Number of Direct Children}
\usage{
plot_number_of_elements_per_folder(x, main = "", to_pdf = FALSE,
  max_chars = 20)
}
\arguments{
\item{x}{tree list structure as returned by \code{\link{to_tree}}}

\item{main}{title of the plot}

\item{to_pdf}{if \code{TRUE} the output is directed to a temporary pdf file}

\item{max_chars}{maximum number of characters to be used for file or folder
names}
}
\description{
Plot Folders with Number of Direct Children
}
