% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treemaps.R
\name{plot_treemaps_from_path_data}
\alias{plot_treemaps_from_path_data}
\title{Plot Treemaps Given File Path Data}
\usage{
plot_treemaps_from_path_data(path_data, pattern = NULL, name = "root",
  as_png = FALSE, n_levels = 3, output_dir = tempdir(),
  type = "value")
}
\arguments{
\item{path_data}{data frame containing file path information as returned by
\code{\link{read_file_info}}}

\item{pattern}{pattern by which to select a subset of paths or \code{NULL}
(default) if all paths in \code{path_data} are to be considered. By
setting the pattern to "^/path/to/start/directory" you can "zoom into" the
treeplot, showing the contents of "/path/to/start/directory".}

\item{name}{name to be used in png file name if \code{as_png} is set. If
\code{path_data} is a list, the names of the list elements are used.}

\item{as_png}{if \code{TRUE} (default) the plots are saved to png files in
the directory given in \code{output_dir} (\code{tempdir()} by default).
Otherwise they are plotted into the active graphical device.}

\item{n_levels}{number of folder depth levels to be shown in the plots}

\item{output_dir}{path to output directory if \code{as_png = TRUE}. Default:
\code{tempdir()}}

\item{type}{passed to \code{\link[treemap]{treemap}}}
}
\description{
Plot Treemaps Given File Path Data
}
