% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treemaps.R
\name{prepare_for_treemap}
\alias{prepare_for_treemap}
\title{Prepare and Filter Path Data for Treemap Plot}
\usage{
prepare_for_treemap(path_data, pattern = NULL, variable = "size",
  path_type = "file", ...)
}
\arguments{
\item{path_data}{data frame as returned by \code{\link{read_file_info}}}

\item{pattern}{regular expression by which to filter the paths}

\item{variable}{name(s) of variable(s) to be selected. Default: "size"}

\item{path_type}{file path_type(s) to filter for. Default: "file"}

\item{...}{further arguments passed to \code{\link{removeCommonRoot}}, such
as \code{n_keep} (number of last segments to be kept from the common first
part of all paths)}
}
\description{
Prepare and Filter Path Data for Treemap Plot
}
