% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_path_data.R
\name{prepare_path_data}
\alias{prepare_path_data}
\title{Select and Filter for Relevant Path Information}
\usage{
prepare_path_data(path_info, pattern = NULL)
}
\arguments{
\item{path_info}{data frame containing file path information as returned by
\code{\link{read_file_info}}}

\item{pattern}{pattern by which to select a subset of paths or \code{NULL}
(default) if all paths in \code{path_info} are to be considered. By
setting the pattern to "^/path/to/start/directory" you can "zoom into" the
path tree, returning only the contents of "/path/to/start/directory".}
}
\value{
data frame with columns \code{path} and \code{size}. See Description.
}
\description{
This function gets a data frame containing path information as input. It
filters for rows with value "file" in column \code{type} and keeps only the
columns \code{path} and \code{size}. If \code{pattern} is not \code{NULL},
the data frame is then filtered for rows in which \code{path} matches the
given pattern. Finally, the common root of all paths in column \code{path} is
removed and the resulting data frame is returned.
}
\examples{
path_info <- kwb.utils::noFactorDataFrame(
  path = c("/path/to/root/", "/path/to/root/file_1", "/path/to/root/file_2"),
  type = c("directory", "file", "file"),
  size = c(0L, 10L, 20L)
)

path_info

kwb.fakin:::prepare_path_data(path_info)
}
