% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree.R
\name{print.path_tree}
\alias{print.path_tree}
\title{Print a tree}
\usage{
\method{print}{path_tree}(tree, max_depth = 2, ...)
}
\arguments{
\item{tree}{object of class \code{"path_tree"} as returned by
\code{kwb.fakin:::to_tree}}

\item{max_depth}{number of depth levels to be printed}

\item{\dots}{further arguments (currently not used)}
}
\description{
Print a tree
}
