% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_properties.R
\name{property_strings_to_data_frame}
\alias{property_strings_to_data_frame}
\title{key1:value1+key2:value2 Strings to Data Frame}
\usage{
property_strings_to_data_frame(property_strings, values = NULL)
}
\arguments{
\item{property_strings}{vector of character containing property values in the
form of \code{prop1:value1+prop2:value2+prob3:value3+...}}

\item{values}{if given and not \code{NULL} this must be a vector of character
as long as \code{property_strings}. The values will appear in a first
column \code{name} of the returned data frame.}
}
\description{
This function needs some improvement! The argument is called "values" but the
values will appear in column "name". Also, what happens if there is also a
property with name "name"?
}
\examples{
property_strings <- c(
  "id:peter+class:person",
  "id:germany+class:country"
)

kwb.fakin:::property_strings_to_data_frame(property_strings)
kwb.fakin:::property_strings_to_data_frame(property_strings, c("a", "b"))

}
\keyword{internal}
