% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_info.R
\name{read_file_info}
\alias{read_file_info}
\title{Deprecated}
\usage{
read_file_info(...)
}
\arguments{
\item{\dots}{arguments passed to private function}
}
\description{
Deprecated
}
\keyword{internal}
