% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_info.R
\name{read_file_info_libuv}
\alias{read_file_info_libuv}
\title{Read CSV File Created Directly With libuv}
\usage{
read_file_info_libuv(file)
}
\arguments{
\item{file}{file containing the output of my own fast version of
\code{fs::dir_info}}
}
\value{
data frame with time columns converted to \code{POSIXct}
}
\description{
Read CSV File Created Directly With libuv
}
