% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_info.R
\name{read_file_info_search_index}
\alias{read_file_info_search_index}
\title{Read CSV File Created With Powershell From Search Index}
\usage{
read_file_info_search_index(file)
}
\arguments{
\item{file}{file containing the output of a Windows powershell script that
reads file paths from the search index database and writes them to a CSV
file}
}
\value{
data frame
}
\description{
Read CSV File Created With Powershell From Search Index
}
