% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_paths.R
\name{read_file_paths}
\alias{read_file_paths}
\title{Read File Paths from a File}
\usage{
read_file_paths(file, metadata = NULL)
}
\arguments{
\item{file}{file containing file path information (path only or additional
information such as file type, size or creation/modification time, etc.)}

\item{metadata}{data frame containing metadata about the file. If given, it
must look as what \code{\link{guess_file_metadata}} returns. If \code{NULL}
the same function is called to guess metadata about the file.}
}
\value{
data frame with columns...
}
\description{
The function tries to guess what type of file is given to the function and
calls the appropriate function to read the file. The aim of this function is
to provide a common result format independent from the type of file that was
read.
}
