% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lines.R
\name{read_lines}
\alias{read_lines}
\title{Read Lines by Giving the File Encoding}
\usage{
read_lines(file, ..., encoding = "unknown", fileEncoding = "")
}
\arguments{
\item{file}{a connection object or character string}

\item{\dots}{arguments passed to \code{\link{readLines}}}

\item{encoding}{passed to \code{\link{readLines}}.}

\item{fileEncoding}{The name of the encoding to be assumed. Passed as
\code{encoding} to \code{\link{file}}, see there.}
}
\description{
Read Lines by Giving the File Encoding
}
