% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressPaths.R
\name{removeCommonRoot}
\alias{removeCommonRoot}
\title{Remove the common root parts}
\usage{
removeCommonRoot(x)
}
\arguments{
\item{x}{list of list of character as returned by
\code{\link[base]{strsplit}}}
}
\description{
Remove the common root parts
}
\examples{
# Split paths at the slashes
absparts <- strsplit(c("a/b/c", "a/b/d", "a/b/e/f/g", "a/b/hi"), "/")

# Remove the common parts of the paths
relparts <- removeCommonRoot(absparts)
relparts

# The extracted root is returned in attribute "root"
attr(relparts, "root")
}
