% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_about_github_package.R
\name{report_about_github_package}
\alias{report_about_github_package}
\title{Report about the Functions in an R-Package on GitHub}
\usage{
report_about_github_package(repo, ...)
}
\arguments{
\item{repo}{path to the GitHub repository}

\item{\dots}{arguments passed to \code{\link{report_about_r_scripts}}}
}
\description{
Report about the Functions in an R-Package on GitHub
}
