% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_about_scripts.R
\name{report_about_r_scripts}
\alias{report_about_r_scripts}
\title{Create a HTML Report About R Scripts}
\usage{
report_about_r_scripts(
  root,
  scripts = dir(root, "\\\\.R$", ignore.case = TRUE, recursive = TRUE),
  show = TRUE
)
}
\arguments{
\item{root}{path to directory from which to start looking for R scripts}

\item{scripts}{optional. Paths to R scripts, relative to the path given in
\code{root}}

\item{show}{if \code{TRUE} the created HTML file is opened in your default
browser}
}
\value{
path to the created HTML file, invisibly
}
\description{
Create a HTML Report about each R script below a root directory. The report
will contain an overview plot showing the number or rows for each script. In
addition, one plot is generated per script, showing for each function
defined in the script, the number of expressions contained in the function.
}
