% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{sortedImportance}
\alias{sortedImportance}
\title{Importance of Strings}
\usage{
sortedImportance(x, weighted = TRUE)
}
\arguments{
\item{x}{vector of character strings}

\item{weighted}{if \code{TRUE} (default) the frequencies of strings are
multiplied by the corresponding string lengths}
}
\value{
named integer vector (of class table) containing the decreasingly
  sorted importance values of the elements in \code{x}. The importance of a
  string is either its frequency in \code{x} (if weighted is FALSE) or the
  product of this frequency and the string length (if weighted is TRUE)
}
\description{
Decreasingly sorted frequencies of strings, by default weighted by their
length
}
\examples{
strings <- c("a", "a", "a", "bc", "bc", "cdefg")

(importance <- kwb.fakin:::sortedImportance(strings))

# Check that each input element is mentioned in the output
all(unique(strings) \%in\% names(importance))

# weighted = FALSE just returns the frequencies of strings in x
(importance <- kwb.fakin:::sortedImportance(strings, weighted = FALSE))

# Check if the sum of frequencies is the number of elements in x
sum(importance) == length(strings)

}
