% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{startsWithParts}
\alias{startsWithParts}
\title{Does a strsplit list start with given sequence of elements?}
\usage{
startsWithParts(parts, elements)
}
\arguments{
\item{parts}{list of list of character as returned by \code{strsplit}}

\item{elements}{vector of character giving the sequence of strings to be
found in \code{parts}}
}
\value{
vector of logical as long as \code{parts} containing \code{TRUE} at
  positions \code{i} for which \code{all(parts[[i]][seq_along(elements)] ==
  elements)} is \code{TRUE}
}
\description{
Does a strsplit list start with given sequence of elements?
}
\examples{
parts <- strsplit(c("a/b/c", "a/b/d", "b/c"), "/")
startsWithParts(parts, c("a", "b"))
startsWithParts(parts, c("b", "c"))
}
