% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.path_tree}
\alias{summary.path_tree}
\title{Get Statistics on Path Tree Nodes}
\usage{
\method{summary}{path_tree}(tree)
}
\arguments{
\item{tree}{object of class \code{path_tree} as returned by
\code{kwb.fakin:::to_tree}}
}
\description{
Get Statistics on Path Tree Nodes
}
