% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{to_tree}
\alias{to_tree}
\title{Convert List of Path Parts to Tree Structure}
\usage{
to_tree(x)
}
\arguments{
\item{x}{list as returned by \code{\link[base]{strsplit}}}
}
\description{
Convert List of Path Parts to Tree Structure
}
