% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{toSubdirMatrix}
\alias{toSubdirMatrix}
\title{Convert a Vector of Paths to a Matrix of Subfolders}
\usage{
toSubdirMatrix(paths, fill.value = "", result_type = "matrix",
  dbg = TRUE)
}
\arguments{
\item{paths}{vector of path strings}

\item{fill.value}{value used to fill empty cells of the result matrix}

\item{result_type}{one of \code{c("matrix", "data.frame")}, specifying the
type of object to be returned}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\value{
matrix or data frame, depending on \code{result_type}
}
\description{
Convert a Vector of Paths to a Matrix of Subfolders
}
\examples{
folder_matrix <- kwb.fakin:::toSubdirMatrix(c("a1/b1/c1", "a1/b2", "a2"))

folder_matrix

dim(folder_matrix)

folder_matrix[folder_matrix[, 1] == "a1", ]

}
