% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_csv}
\alias{write_csv}
\title{Write Data Frame to CSV File}
\usage{
write_csv(data, file, sep = ";", version = 2, ...)
}
\arguments{
\item{data}{data frame}

\item{file}{path to CSV file to be written}

\item{sep}{column separator}

\item{version}{determines which function to use for writing the CSV file
1: \code{\link[utils]{write.table}}, 2: \code{\link[data.table]{fwrite}}}

\item{\dots}{further arguments passed to \code{\link[utils]{write.table}} or
\code{\link[data.table]{fwrite}}}
}
\description{
Write Data Frame to CSV File
}
