% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_file_info.R
\name{write_file_info}
\alias{write_file_info}
\title{Write File Information to CSV File}
\usage{
write_file_info(file_info, file, version = 2)
}
\arguments{
\item{file_info}{data frame as returned by
\code{\link{get_recursive_file_info}}}

\item{file}{path to CSV file to be written}

\item{version}{determines which function to use for writing the CSV file
1: \code{\link[utils]{write.table}}, 2: \code{\link[data.table]{fwrite}}}
}
\description{
Write File Information to CSV File
}
