% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialChars.R
\name{get_special_character_info}
\alias{get_special_character_info}
\title{Get Special Characters and Their Byte Codes}
\usage{
get_special_character_info(text, context_length = 7, bytes_per_char = 2)
}
\arguments{
\item{text}{vector of character of length one}

\item{context_length}{number of characters left and right of special
character to be put into column \code{context}}

\item{bytes_per_char}{number of bytes per character}
}
\value{
data frame with columns \code{special} (special characters) and
  \code{bytes} (hexadecimanl byte codes as a space separated string),
  \code{context} (strings "around" the special characters)
}
\description{
Get Special Characters and Their Byte Codes
}
\examples{
(text <- kwb.fakin:::example_string_with_specials("de"))

get_special_character_info(text)

}
