% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{getProjectPaths}
\alias{getProjectPaths}
\title{Paths where to find project on the KWB server}
\usage{
getProjectPaths(start_directory, as_list = FALSE)
}
\arguments{
\item{start_directory}{Path to the project network drive at KWB}

\item{as_list}{If \code{TRUE} (the default is \code{FALSE}) the paths
are returned as a list with the folder names as list element names.}
}
\value{
full paths to project folders as a vector of character or as a named
  list if \code{as_list = TRUE}.
}
\description{
Paths where to find project on the KWB server
}
