% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey-network.R
\name{plot_path_network}
\alias{plot_path_network}
\title{Plot Paths as Sankey Network}
\usage{
plot_path_network(paths, max_depth = 2, ...)
}
\arguments{
\item{paths}{character vector of paths}

\item{max_depth}{maximum depth of paths to be shown}

\item{\dots}{arguments passed to \code{\link[networkD3]{sankeyNetwork}}, such
as \code{nodeWidth}, \code{nodePadding}, \code{fontSize}}
}
\value{
object representing an HTML page
}
\description{
Plot Paths as Sankey Network
}
\examples{
# Get the paths to all folders on the desktop
paths <- dir(system.file(package = "kwb.fakin"), recursive = TRUE)

# Plot the folder network
plot_path_network(paths)

}
