% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path_network.R
\name{plot_path_network}
\alias{plot_path_network}
\title{Plot Paths as Sankey Network}
\usage{
plot_path_network(paths, max_depth = 3, nodePadding = 8,
  nodeHeight = 10, sinksRight = FALSE, remove_common_root = TRUE,
  names_to_colours = name_to_traffic_light, height = NULL, ...,
  method = 1, weight_by = c("n_files", "size", "none")[1],
  sizes = NULL)
}
\arguments{
\item{paths}{character vector of paths}

\item{max_depth}{maximum depth of paths to be shown}

\item{nodePadding}{passed to \code{\link[networkD3]{sankeyNetwork}}, see
there. Decrease this value (e.g. `nodePadding = 0`) if there are
many nodes to plot and the plot does not look as expected}

\item{nodeHeight}{height of a node in pixels. Used to calculate the total
plot height.}

\item{sinksRight}{passed to \code{\link[networkD3]{sankeyNetwork}}, see there}

\item{remove_common_root}{remove the common root parts? (default: TRUE)}

\item{names_to_colours}{if not \code{NULL} expected to be a function that
accepts a vector of (node) names and returns a vector of (colour) names
of same length. This function will be called by \code{plot_path_network}
to determine the colour for each node based on its name. By default, the
function \code{\link{name_to_traffic_light}} is called.}

\item{height}{plot height in pixels, passed to
\code{\link[networkD3]{sankeyNetwork}}. If \code{NULL}, the height is
calculated based on \code{nodeHeight}, \code{nodePadding} and the maximum
number of nodes at one folder depth.}

\item{\dots}{further arguments passed to
\code{\link[networkD3]{sankeyNetwork}}, such as \code{nodeWidth},
\code{nodePadding}, \code{fontSize}}

\item{method}{if \code{1} (default) the function behaves as before, another
value activates the new preparation of paths accepting/using an object
of class \pkg{pathlist}}

\item{weight_by}{one of \code{"n_files", "size", "none"}. Specifies whether
to set the link widths according to the total number or total size of files
in subsequent folders or by setting all links to the same width.}

\item{sizes}{file sizes corresponding to the \code{paths}}
}
\value{
object representing an HTML page
}
\description{
Plot Paths as Sankey Network
}
\examples{
# Get the paths to all folders on the desktop
paths <- dir(system.file(package = "kwb.fakin"), recursive = TRUE)

# Plot the folder network
plot_path_network(paths)

}
