% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treemaps.R
\name{plot_treemaps_from_path_data}
\alias{plot_treemaps_from_path_data}
\title{Plot Treemaps Given File Path Data}
\usage{
plot_treemaps_from_path_data(path_data, root_path = "", name = "root",
  as_png = FALSE, n_levels = 3, output_dir = tempdir(),
  type = "value", args_png = list(), n_biggest = -1, depth = 1)
}
\arguments{
\item{path_data}{data frame containing file path information as returned by
\code{\link{read_file_info}}}

\item{root_path}{path to the folder that contains all paths to be considered.
By setting the root path to "/path/to/root" you can "zoom into" the
treeplot, showing the contents below "/path/to/root" only. If
\code{root_path} is \code{""} (default) all paths in \code{path_data} are
considered.}

\item{name}{name to be used in png file name if \code{as_png} is set. If
\code{path_data} is a list, the names of the list elements are used.}

\item{as_png}{if \code{TRUE} (default) the plots are saved to png files in
the directory given in \code{output_dir} (\code{tempdir()} by default).
Otherwise they are plotted into the active graphical device.}

\item{n_levels}{number of folder depth levels to be shown in the plots}

\item{output_dir}{path to output directory if \code{as_png = TRUE}. Default:
\code{tempdir()}}

\item{type}{passed to \code{\link[treemap]{treemap}}}

\item{args_png}{list of arguments passed to \code{\link[grDevices]{png}} if
\code{as_png = TRUE}}

\item{n_biggest}{vector of integer, specifying the number(s) of biggest
folders (in terms of size and number of files) in which to "zoom into". The
position in the vector represents the folder depth. For example, if
\code{n_biggest = c(2, 1)}, the first element (\code{1}) indicates that
sub-treemaps are produced for the two biggest subfolders below
\code{root_path}: for \code{root_path/biggest-1} and for
\code{root_path/biggest-2}. The second element (\code{1}) indicates that
further treemaps are generated only for the biggest subfolders below
\code{root_path/biggest-1} and \code{root_path/biggest-2}, respectively,
in each case. The length of the vector \code{n_biggest} determines the
maximal depth until which to generate treemaps. By setting an element
to \code{-1L} you specify that sub-treemaps are generated for each
subfolder on the corresponding folder depth.}

\item{depth}{current depth of recursion}
}
\description{
Plot Treemaps Given File Path Data
}
