% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_path_tree.R
\name{print.path_tree}
\alias{print.path_tree}
\title{Print a tree}
\usage{
\method{print}{path_tree}(x, max_depth = 2, ...)
}
\arguments{
\item{x}{tree object as returned by \code{kwb.fakin:::to_tree}}

\item{max_depth}{number of depth levels to be printed}

\item{\dots}{further arguments (currently not used)}
}
\description{
Print a tree
}
