% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_file_info.R
\name{read_file_info}
\alias{read_file_info}
\title{Read File Information from CSV File}
\usage{
read_file_info(file, version = 2, ...)
}
\arguments{
\item{file}{path to CSV file containing file information as stored with
\code{\link{write_file_info}}}

\item{version}{passed to \code{\link{read_csv}}}

\item{\dots}{further arguments passed to \code{\link{read_csv}})}
}
\description{
Read File Information from CSV File
}
