% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_path_information.R
\name{read_path_information}
\alias{read_path_information}
\title{Read Files Containing File Path Information from Folder}
\usage{
read_path_information(file_info_dir, pattern = "^path-info", sep = ",",
  ...)
}
\arguments{
\item{file_info_dir}{path to folder in which to look for files matching
\code{pattern}}

\item{pattern}{pattern to match against the file names to be read. Default:
"^path-info"}

\item{sep}{column separator passed to \code{\link{read_file_info}}}

\item{\dots}{further arguments passed to \code{\link{read_file_info}}}
}
\description{
Read Files Containing File Path Information from Folder
}
\examples{
# Set root directory (here: package installation directory of kwb.fakin)
root_dir <- system.file(package = "kwb.fakin")

# Set output directory
output_dir <- tempdir()

# Write all paths below root_dir into a "path-info"-file
kwb.fakin::get_and_save_file_info(root_dir, output_dir)

# Read the "path-info"-files that are (now) found in output_dir
path_info <- kwb.fakin:::read_path_information(output_dir, sep = ";")

}
