% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_paths}
\alias{read_paths}
\title{Read Paths From File}
\usage{
read_paths(file, encoding = NULL, do_sort = TRUE)
}
\arguments{
\item{file}{full path to the text file containing paths}

\item{encoding}{passed to \code{\link[base]{readLines}}}

\item{do_sort}{if \code{TRUE} (default), the vector of paths is sorted
alphanumerically.}
}
\value{
vector of character
}
\description{
Read paths from file, convert backslashes to slashes and sort the paths
(by default)
}
