% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{read_paths}
\alias{read_paths}
\title{Read Paths From File}
\usage{
read_paths(file, ..., do_sort = TRUE)
}
\arguments{
\item{file}{full path to the text file containing paths}

\item{\dots}{arguments passed to \code{\link{read_lines}}, such as
\code{n}, \code{fileEncoding}, \code{encoding}}

\item{do_sort}{if \code{TRUE} (default), the vector of paths is sorted
alphanumerically.}
}
\value{
vector of character
}
\description{
Read paths from a file that contains one path per line. Convert backslashes
to slashes and sort the paths (by default)
}
