% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_path.R
\name{removeCommonRoot}
\alias{removeCommonRoot}
\title{Remove the common root parts}
\usage{
removeCommonRoot(x, n_keep = 0, dbg = TRUE)
}
\arguments{
\item{x}{list of vectors of character as returned by
\code{\link[base]{strsplit}} or a vector of character.}

\item{n_keep}{number of common path segments to keep (so that the path
tree keeps its root)}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\description{
Remove the common root parts
}
\examples{
# Split paths at the slashes
absparts <- strsplit(c("a/b/c", "a/b/d", "a/b/e/f/g", "a/b/hi"), "/")

# Remove the common parts of the paths
relparts <- removeCommonRoot(absparts)
relparts

# The extracted root is returned in attribute "root"
attr(relparts, "root")

}
