% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.path_tree}
\alias{summary.path_tree}
\title{Get Statistics on Path Tree Nodes}
\usage{
\method{summary}{path_tree}(object, ...)
}
\arguments{
\item{object}{tree object as returned by \code{kwb.fakin:::to_tree}}

\item{\dots}{further arguments (currently not used)}
}
\description{
Get Statistics on Path Tree Nodes
}
